Workaround kernel bug which does not respect `setaffinity`
There is a kernel bug in the kernels we use on our phones wherein the affinity
mask set by `sched_setaffinity` is cleared when a process is migrated to a
different cpuset (even if the two cpusets have the exact same cores allocated
to them).

This bug was resulting in us throwing out many results due to the benchmark
swapping between two different cpusets right at the end of the benchmark
execution, invalidating that run.

See here for the fix in the kernel [1, 2].
[1]: https://lore.kernel.org/all/20220922180041.1768141-3-longman@redhat.com/T/#m51982911c50a294c26f9f84f999081d963250a86
[2]: https://github.com/torvalds/linux/commit/8f9ea86fdf99b81458cc21fc1c591fcd4a0fa1f4